app/requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
psycopg2-binary>=2.9.0
pyjwt>=2.8.0
requests>=2.31.0
lxml>=4.9.0
